+++ Changelog.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file since v2.0.0
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.0] - 2022-??-??
+## [3.1.0] - 2022-09-05
 
 ### Added
 
 - Added and tested Magento 2.4.5 templates.
 - Slightly better error handling for `composer create-project`.
+- Added more tests to the `magento:test-dockerfiles` command to cover more issues.
+- Implemented simple check for xDebug configuration in `magento:test-dockerfiles`.
 
 ### Changed